Auth Finished
Finished the Auth Page. Now, when a user successfully signs back in.
This will trigger our authChannel which tells our RootWindow to create
and to push the rest of TermLink's Pages onto the application.

Started and built the Header and a BulletItem UI element which wraps
an Item in Half Circles, giving it the appearance of rounded corners.
Though.. this doesn' work with all colors.