Contacts Page
Created the Contacts Page -- Currently displays the Contacts in a
ScrollView that I created. At the moment, when you select a Contact,
this updates the ContactBody with the selected Contact.

What I need to do next is design the UI for the ContactPage's Body.
Which will show
 - Data about the Contact
 - Current Messages between the Contact
 - Actions that the user can perform with said Contact
 - and other information that I haven't though of yet

Another thing I'll have to develop is the Data Layer for how we
initially load in the contacts data( at the moment I've created a
temporary map of *ContactItems for development)

A Few things that'll need implemented
 - Initially loading in the data on ContactPage load
 - Building and running a background Go Routine that
   will periodically check for any Contact Updates(i.e., Status updates)

At the moment, these implementations should be trivial for where
ContactBody is currently at presently